[CHANGE] rename etag_helper
