+++ ledger/helpers/etag.py
